+++ packages/fdc3-agent-proxy/src/DesktopAgentProxy.ts
@@ -70,6 +70,7 @@ export class DesktopAgentProxy implements DesktopAgent, Connectable {
     this.addIntentListenerWithContext = this.addIntentListenerWithContext.bind(this);
     this.raiseIntentForContext = this.raiseIntentForContext.bind(this);
     this.open = this.open.bind(this);
+    this.close = this.close.bind(this);
     this.findInstances = this.findInstances.bind(this);
     this.getAppMetadata = this.getAppMetadata.bind(this);
     this.disconnect = this.disconnect.bind(this);
@@ -198,6 +199,10 @@ export class DesktopAgentProxy implements DesktopAgent, Connectable {
 
   open(app: string | AppIdentifier, context?: Context | null, metadata?: AppProvidableContextMetadata) {
     return this.apps.open(this.ensureAppId(app)!, context, metadata);
+  }
+
+  close(): Promise<void> {
+    return this.apps.close();
   }
 
   findInstances(app: AppIdentifier) {